change `size_t` to `uint32_t`

+++ src/common.h
@@ -4,14 +4,14 @@
 #define VERSION "0.0.1"
 
 #include <stdlib.h>
-
+#include <stdint.h>
 #define TRUE 1
 #define FALSE 0
 
 
 
-void * xmalloc(size_t size);
-void * xrealloc(size_t size);
-void * tab_print(size_t size);
+void * xmalloc(uint32_t size);
+void * xrealloc(uint32_t size);
+void * tab_print(uint32_t size);
 
 #endif
